requirements.txt
# 스케줄링 및 작업 큐
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Core Data Engineering
kafka-python==2.0.2
confluent-kafka==2.6.1
pyspark==3.4.3
py4j==0.10.9.7

# Database & Storage
psycopg2-binary==2.9.9
influxdb-client==1.43.0
sqlalchemy==2.0.25

# Data Processing
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.33
requests==2.32.3

# Monitoring & Data Quality
prometheus-client==0.19.0
grafana-api==1.0.3
great-expectations==0.18.19

# API & Web Framework
fastapi==0.112.1
uvicorn==0.30.6
pydantic==2.8.2

# Utilities and Logging
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.0
schedule==1.2.2
aiofiles==24.1.0

# Jupyter 및 개발환경(선택)
notebook
jupyterlab
ipykernel
